[py] Allow user to set max. # of redirects
Allow user to set max. # of redirects

Selenium currently does not allow to set the number of redirects allowed for a request through urrlib3. When the number of retries, and redirects is not specified, urrlib3 defaults to 3 retries, and 3 redirects. This can pose an issue for automation testers that rely on third-party servers to perform their tests, as the server could be busy when selenium is attempting to redirect. Giving the option to customize the number of redirects can be very beneficial for automation devs, to provide flexibility in other clients they wish to use.